+++ src/application/VulcanBuild.kt
@@ -1,5 +1,6 @@
 package application
 
+import builder.BlockBuilder
 import builder.ItemBuilder
 import builder.ModBuilder
 import io.Directories
@@ -82,6 +83,7 @@ object VulcanBuild {
 
         when(type) {
             "item" -> ItemBuilder(fileName, lines).build()
+            "block" -> BlockBuilder(fileName, lines).build()
             else -> UIHandler.error("Unrecognised type for $fileName: \"$type\". This file will be skipped.")
         }
     }